+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.13.2
-- hide UUID col in table
-- noEdit not evaluated for create state entity form
-- delete state transition
+- Option to hide UUID col in table
+- State model: export/import
+- State model: delete transition
+- Fix noEdit not evaluated for create state entity form
 
 # 0.13.1
 - fix multi entity create